Added missing javadocs and moved the base code to the core library
