+++ app/src/main/java/com/xmartlabs/template/ApplicationComponent.java
@@ -1,16 +1,15 @@
 package com.xmartlabs.template;
 
 import com.xmartlabs.base.core.CoreComponent;
+import com.xmartlabs.base.core.controller.Controller;
 import com.xmartlabs.base.core.module.CoreAndroidModule;
 import com.xmartlabs.base.core.module.GsonModule;
 import com.xmartlabs.base.core.module.LoggerModule;
 import com.xmartlabs.base.core.module.OkHttpModule;
 import com.xmartlabs.base.core.module.PicassoModule;
 import com.xmartlabs.base.core.module.RestServiceModule;
-import com.xmartlabs.template.controller.Controller;
-import com.xmartlabs.template.controller.ServiceController;
 import com.xmartlabs.template.controller.SessionController;
-import com.xmartlabs.template.controller.SharedPreferencesController;
+import com.xmartlabs.base.core.controller.SharedPreferencesController;
 import com.xmartlabs.template.module.ControllerModule;
 import com.xmartlabs.template.ui.BaseActivity;
 import com.xmartlabs.template.ui.BaseAppCompatActivity;
