+++ app/src/main/java/com/xmartlabs/template/module/ControllerModule.java
@@ -1,7 +1,7 @@
 package com.xmartlabs.template.module;
 
 import com.xmartlabs.template.controller.SessionController;
-import com.xmartlabs.template.controller.SharedPreferencesController;
+import com.xmartlabs.base.core.controller.SharedPreferencesController;
 
 import javax.inject.Singleton;
 
@@ -14,11 +14,5 @@ public class ControllerModule {
   @Singleton
   SessionController provideSessionController() {
     return new SessionController();
-  }
-
-  @Provides
-  @Singleton
-  SharedPreferencesController provideSharedPreferencesController() {
-    return new SharedPreferencesController();
   }
 }
