+++ core/build.gradle
@@ -183,6 +183,8 @@ dependencies {
     compile "com.squareup.okhttp3:okhttp:${project.ext.okHttpVersion}"
     compile "com.squareup.okhttp3:logging-interceptor:${project.ext.okHttpVersion}"
     compile 'com.wdullaer:materialdatetimepicker:2.5.0'
+    compile "io.reactivex.rxjava2:rxandroid:${project.ext.rxAndroidVersion}"
+    compile "io.reactivex.rxjava2:rxjava:${project.ext.rxJavaVersion}"
     provided "com.google.dagger:dagger-compiler:${project.ext.daggerVersion}"
     provided "org.projectlombok:lombok:${project.ext.lombokVersion}"
     testCompile 'junit:junit:4.12'
