+++ core/src/main/java/com/xmartlabs/base/core/model/EntityWithId.java
@@ -1,4 +1,4 @@
-package com.xmartlabs.template.model;
+package com.xmartlabs.base.core.model;
 
 public interface EntityWithId<T> {
   T getId();
